Changed data type of weight and rpe

+++ Model/Lift.cs
@@ -4,8 +4,8 @@ namespace GymLifts.Model;
 
 public class Lift
 {
-    public float Weight { get; set; }
+    public double Weight { get; set; }
     public int Reps { get; set; }
-    public float RPE { get; set; }
+    public double RPE { get; set; }
     public string Time { get; set; }
 }
